Exception handling - delete

+++ src/main/java/com/studies/course/services/UserService.java
@@ -14,26 +14,31 @@ import com.studies.course.services.exceptions.ResourceNotFoundException;
 public class UserService {
 	@Autowired
 	private UserRepository repository;
-	
+
 	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
-	}
-	
+		if (repository.existsById(id)) {
+			repository.deleteById(id);
+		}
+		 else {
+			throw new ResourceNotFoundException(id);				
+		}
+	} 
+
 	public User update(Long id, User obj) {
-		//User entity = repository.getReferenceById(id);
+		// User entity = repository.getReferenceById(id);
 		User entity = findById(id);
 		updateData(entity, obj);
 		return repository.save(entity);
@@ -43,7 +48,7 @@ public class UserService {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
-		
+
 	}
-	
+
 }
